Updated dependencies
Updated numerous dependencies including upgrading to Gradle plugin v2

Made some changes to the db persistence services as an exception was
being thrown upon application startup that was squelched with
RxJava/RxAndroid 1.1.0/1.1.0 but caused all DB get operations to fail with
versions 1.1.5/1.2.0. The new approach takes a sample() of the
observable before going through onBackpressureLatest, which seems to
work although may need revisiting in the future.